lasr/cpython/main.py
import sys
from dataclasses import dataclass


# Press ⇧F10 to execute it or replace it with your code.
# Press Double ⇧ to search everywhere for classes, files, tool windows, actions, and settings.


def crack() -> int:
    result : int = 42
    return result


@dataclass
class StringIO:
    _buf     : str
    _0cursor : int = int(0)
    _len     : int = int(0)


def __post_init__(this: StringIO):
    this._len = len(this._buf)


def read(this: StringIO) -> str:
    result = ''
    return result


if __name__ == '__main__':
    integer_asr : str = '(Integer 4 [])'
    test_dude   : StringIO = StringIO(integer_asr, 0, 0)
    __post_init__(test_dude)  # Manual Call! SHOULD NOT NEED THIS
    assert test_dude._buf == integer_asr
    assert test_dude._len == len(integer_asr)


